requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
email-validator>=2.1.0
motor>=3.3.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
APScheduler>=3.10.4
watchtower>=3.0.1
boto3>=1.28.44
python-dotenv>=1.0.0
google-auth>=2.23.3
google-auth-oauthlib>=1.1.0